chore: crypto contacts code style (#10523)
